+++ init.js
@@ -1,0 +1,8 @@
+import {loadToDos} from "./todoApp.js"
+function init() {
+    console.log("load toDos");
+    loadToDos();
+    
+}
+
+init();
