+++ todoApp.js
@@ -1,22 +1,14 @@
-import  { handleCount, viewActiveClick, viewAllClick, viewCompletedClick,completedEl,allEl,filterState, renderActiveItems } from "./component/todoCount.js";
-
-
-
-export let toDos = [
-
-];
-
-const toDoInput = document.getElementById("new-todo-title");
+import  { handleCount, viewCompletedClick,completedEl, filterState, renderActiveItems, renderCompleteItems } from "./component/todoCount.js";
+import {handleTodoItemClick} from "./component/todoList.js"
+import "./component/todoList.js";
 export const todoListEl = document.getElementById('todo-list');
 
-
-
-
+const toDoInput = document.getElementById("new-todo-title");
 const labelEl = document.querySelector(".label");
-const eidtInputEl = document.querySelectorAll(".edit")
-
 
 const TODOS_LS ="toDos";
+export let toDos = [];
+
 
 export const addToDos = (item)=>{
     todoListEl.insertAdjacentHTML("beforeend",renderTodoItemTemplate(item));
@@ -25,13 +17,6 @@ export const addToDos = (item)=>{
 
 const addToItems = (item)=>{
     toDos = [item, ...toDos];   
-    saveToDos();
-}
-
-
-const removeFromItems=(li)=>{
-    const testItemId = li.dataset.id;
-    toDos = toDos.filter(item => `${item.id}` !== testItemId);
     saveToDos();
 }
 
@@ -55,46 +40,8 @@ const handleNewTodoSubmit = async (event)=>{
 };
 
 
-const itemsUpdate=(event)=>{
-
-    const currentLi = event.target.closest('li');
-
-    const currentItemId = currentLi.dataset.id;
-    for(let obj of toDos){
-
-        if(obj.completed === false && obj.id === parseInt(currentItemId)){
-            obj.completed = true;
-            
-        } else if(obj.completed === true && obj.id === parseInt(currentItemId)){
-            obj.completed = false;
-        }
-    }
-    if(filterState === "completed"){
-        renderCompleteItems();
-    } else if (filterState === "active"){
-        renderActiveItems();
-    }
-
-}
-    
-const handleComplete=(event)=>{
-    event.target.closest("li").classList.toggle("completed");
-    event.target.closest("input").classList.toggle("checked");
-    itemsUpdate(event);
-}
-
-
-const handleDestory=(event)=>{
-    const li = event.target.parentNode.parentNode;
-    removeFromItems(li);
-    todoListEl.removeChild(li);
-
-    handleCount(toDos.length);
-}
-
 const updateEditTitle=(event)=>{
     for(let obj of toDos){
-
         if( obj.id === parseInt(event.path[1].dataset.id)){
             obj.title = event.path[0].value;
         }
@@ -108,7 +55,6 @@ const updateEdit=(event)=>{
 
         event.path[1].childNodes[1].childNodes[3].innerText=event.path[0].value;
         event.path[1].classList.remove('editing');
-
         updateEditTitle(event);
         
     } else if(event.keyCode === 27){
@@ -118,7 +64,7 @@ const updateEdit=(event)=>{
     
 }
 const handleEdinting=async(event)=>{
-    const targetInput =event.target.parentNode.nextSibling.nextSibling
+    const targetInput =event.target.parentNode.nextSibling.nextSibling;
     
     try{
         await targetInput.addEventListener("keyup",updateEdit);
@@ -134,20 +80,14 @@ const handleEdit=(event)=>{
     const targetLi = event.target.closest('li');
     targetLi.classList.add("editing");
     handleEdinting(event);
-    ;
-}
-
-const handleTodoItemClick=(event)=>{
-    const targetClass = event.target.className.split(" ");
-    
-    if(targetClass[0] === "toggle") handleComplete(event);
-    else if(targetClass[0] === "destroy") handleDestory(event);
 
 }
 
 
 
-const renderTodoItemTemplate=(item)=>{
+
+
+export const renderTodoItemTemplate=(item)=>{
     return (
      `<li data-id="${item.id}" class="${item.completed ? "completed" : ""}">
         <div class="view">
@@ -160,14 +100,11 @@ const renderTodoItemTemplate=(item)=>{
 };
 
 
-
-
-
-
-const saveToDos=()=>{
+export const saveToDos=()=>{
     localStorage.setItem(TODOS_LS,JSON.stringify(toDos));
 }
-const loadToDos=()=>{
+
+export const loadToDos=()=>{
 
     const loadedToDos = localStorage.getItem(TODOS_LS);
 
@@ -179,21 +116,13 @@ const loadToDos=()=>{
             addToItems(toDo)
         });
     }
-}
-
-function init() {
-    console.log("load toDos");
-    loadToDos();
-
-    todoListEl && todoListEl.addEventListener("click",handleTodoItemClick);
-    todoListEl && todoListEl.addEventListener("dblclick",handleEdit);
-    toDoInput && toDoInput.addEventListener( "change",handleNewTodoSubmit); 
-    allEl && allEl.addEventListener("click",viewAllClick);
-
-    completedEl && completedEl.addEventListener("click",viewCompletedClick);
-    labelEl && labelEl.addEventListener("click",handleEdit);
-
     handleCount(toDos.length);
 }
 
-init();
+function todoApp(){
+    todoListEl && todoListEl.addEventListener("click",handleTodoItemClick);
+    labelEl && labelEl.addEventListener("click",handleEdit);
+    todoListEl && todoListEl.addEventListener("dblclick",handleEdit);
+    toDoInput && toDoInput.addEventListener( "change",handleNewTodoSubmit); 
+}
+todoApp();
